pyproject.toml
[tool.poetry]
name = "igvf"
version = "0.1.0"
description = ""
authors = ["Daniel Zeiberg <zeiberg.d@northeastern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lmdb = "^1.4.1"
lmdbm = "^0.0.5"
ipykernel = "^6.29.3"
mysql-connector-python = "^8.3.0"
biopython = "^1.83"
tqdm = "^4.66.2"
pandas = "^2.2.1"
fire = "^0.6.0"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
scipy = "^1.12.0"
sqlalchemy = "^2.0.29"
voila = "^0.5.5"
pip = "^24.0"
install = "^1.3.5"
ipywidgets = "^8.1.2"
viola = "^0.3.8"
streamlit = "^1.32.2"
watchdog = "^4.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
